This is a monster checkin that really should have been done a lot before, but many of the changes are dependent on each other.  This checkin is an intermediate version that is not complete and may not even compile, but I don't dare sit on the changes any longer.
One of the major changes is the API for the state machine system which
has been changed a lot to facilitate some new features.  The most obvious
result of this is that some of the code in G3/fsm.c now needs to be
modified.  I hope this is a good change, although the syntax is
doubious as it makes heavy use of the C preprocessor and the state
machine functions really don't look like C functions any more.

Another important change is the HAL, which should make it easier to
make this code work on non-ISDN hardware.

Take care; much work remains.